# openapi-user-content.json
{
  "openapi": "3.1.0",
  "info": { "title": "OBPC User Content API", "version": "1.1.0" },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
...
  },
  "paths": {
    "/api/admin/uploads": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object", "required": ["file", "context"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "context": { "type": "string", "enum": ["ministerios", "eventos", "noticias", "mensagens", "louvores"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "URL HTTPS publica" }, "400": { "description": "Arquivo invalido" }, "413": { "description": "Arquivo maior que 8 MB" } }
      }
    },
...
  }
}